[WIP] Hash Cache
As the hash table inflates our query times have gone from 290us
to like 2.2ms. This is a growing rate. We can't eager load, because our
table has no foreign local key pair, its simply some attributes in
the host table need to be eager loaded to hashes.

To fix this, we eager load accounts/characters and nothing else. We
then calculate all the hashes we WILL need to load, but haven't. We
then run that query using whereIn() to grab them all and cache them.

That page load then has all the hashes it needs without having to do
Hashes::all().

This exponential growth of load times can be shown by the 589%
increase in load time for the queries. This change for roster/profile
brings it back down to 490us. Which is quite a drop in time.